requirements.txt
# Clinical-QA-Sentinel Requirements
# Updated with refactoring dependencies

# ============================================================================
# Core Testing Framework
# ============================================================================
selenium>=4.16.0
webdriver-manager>=4.0.1

# ============================================================================
# Test Data Generation
# ============================================================================
faker>=22.0.0

# ============================================================================
# Configuration Management
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# Testing & Quality Assurance
# ============================================================================
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-cov>=4.1.0

# ============================================================================
# Utilities
# ============================================================================
python-dateutil>=2.8.2

# ============================================================================
# Optional: Advanced Features
# ============================================================================
# requests>=2.31.0
# pandas>=2.1.0
# openpyxl>=3.1.2

# ============================================================================
# Development Dependencies
# ============================================================================
# autopep8>=2.0.4
# pylint>=3.0.3
# black>=23.12.0
